feat(file-transfer): improve UX
Show bucket info during confirm; improve progress bar UI, prevent inconsistent width
