+++ packages/tauri-plugins/jarvis/permissions/all.toml
@@ -94,7 +94,6 @@ commands.allow = [
     "get_file_transfer_bucket_keys",
     "get_file_transfer_bucket_by_key",
     "local_net_send_file",
-    "download_file",
     "download_files",
     "file_transfer_preview_bucket",
 ]
