+++ packages/tauri-plugins/jarvis/src/lib.rs
@@ -177,12 +177,11 @@ pub fn init<R: Runtime>(db_key: Option<String>) -> TauriPlugin<R> {
             /*                                File Transfer                               */
             /* -------------------------------------------------------------------------- */
             // commands::server::get_files_to_send,
-            commands::server::get_file_transfer_bucket_keys,
-            commands::server::get_file_transfer_bucket_by_key,
-            commands::server::local_net_send_file,
-            commands::server::download_file,
+            commands::file_transfer::get_file_transfer_bucket_keys,
+            commands::file_transfer::get_file_transfer_bucket_by_key,
+            commands::file_transfer::local_net_send_file,
             commands::file_transfer::download_files,
-            commands::server::file_transfer_preview_bucket,
+            commands::file_transfer::file_transfer_preview_bucket,
         ])
         .setup(move |app, api| {
             utils::setup::setup_app_path(app);
